src/requirements.txt
email_validator==1.1.0
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Classful==0.14.1
Flask-Login==0.4.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
pytest==5.4.1
pytest-flask==1.0.0
python-dotenv==0.13.0
requests==2.21.0
validators==0.15.0
